feat: integrate Clerk for authentication and user management

+++ src/controllers/cron-controllers/delete-cron-controller.ts
@@ -1,17 +1,29 @@
-import { NextFunction, Request, Response } from 'express';
+import { Request, Response } from 'express';
 import { asyncHandler } from '../../utils/async-handler';
 import { prisma } from '../../lib/prisma';
 import APIResponseType from '../../types/response.type';
 import { stopCronJob } from '../../services/cron-scheduler.service';
 
 const deleteCronController = asyncHandler(
-  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-    const { id } = req.params;
+  async (req: Request, res: Response): Promise<void> => {
 
-    const cronBeforeDelete = await prisma.cron.findUnique({
-      where: { id },
+    const { id } = req.params;
+    const user = req.user;
+
+    const cronBeforeDelete = await prisma.cron.findFirst({
+      where: { id, userId: user.id },
       include: { url: true },
     });
+
+    if (!cronBeforeDelete) {
+      res.status(404).json({
+        success: false,
+        message: 'Cron not found or unauthorized',
+        data: null,
+        statusCode: 404,
+      } as APIResponseType);
+      return;
+    }
 
     const deletedCron = await prisma.cron.delete({
       where: {
